[feat] 最新の記録表示機能をTrainingStatsからNextGoalSelectorに移動
- NextGoalSelector に最新の記録を表示する機能を実装
- TrainingStats から最新の記録表示部分を削除
- 各トレーニングタブのコンポーネント配置を微調整
- レスポンシブデザインの細かいスタイル調整
